services/text_embedding/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
requests
